[backup] refactoring backup
Position validate refactoring
