+++ src/main/java/Numbers/Number.java
@@ -1,0 +1,5 @@
+package Numbers;
+
+public interface Number {
+    public int get();
+}
