+++ src/main/java/ladder/Node.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import Numbers.Position;
+
 public class Node {
     final Direction Dir;
