+++ src/test/java/ladder/LadderCreatorTest.java
@@ -1,4 +1,5 @@
 package ladder;
+import Numbers.NaturalNum;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
