+++ src/test/java/ladder/RowTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import Numbers.NaturalNum;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
